Time: 0 ms (100.00%), Space: 5.9 MB (96.30%) - LeetHub

+++ 0070-climbing-stairs/0070-climbing-stairs.cpp
@@ -1,11 +1,11 @@
 class Solution {
 public:
     int climbStairs(int n) {
-        long long o = 0;
-        long long s = 1;
+        int o = 0;
+        int s = 1;
         while(n-- > 0)
         {
-            long long t = o + s;
+            int t = o + s;
             o = s;
             s = t;
         }
